% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCloud.R
\name{checkCloud}
\alias{checkCloud}
\title{Check Cloud function}
\usage{
checkCloud(dbCloud, whichDay = Sys.Date() - 1)
}
\arguments{
\item{dbCloud}{The CloudDB returned by ReadServerData()}

\item{whichDay}{The day to filter data and produce reports? Defaults to Sys.Date()-1}
}
\value{
A list with the DB filtered on the selected wichDay and the alert DB with detected anomalies registered.
}
\description{
Finds anomalies and prepare data for reports.
}
\examples{
checkTT(dbCloud, Sys.Date()-1)
}
\keyword{TreeTalkers}
