% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRequirements.R
\name{checkRequirements}
\alias{checkRequirements}
\title{Check requirements function}
\usage{
checkRequirements(whichDay = Sys.Date() - 1)
}
\arguments{
\item{whichDay}{The day to filter data and produce reports? Defaults to Sys.Date()-1}
}
\value{
This function saves a file: file.path(tempdir(), "whichDay.txt")
}
\description{
Prepare data and handle any exception that can occur each time a new user runs the TreeTalkersCheck package for the first time. Also saves a "whichDay.txt" temporary file.
}
\examples{
checkRequirements()
checkRequirements("2020-09-16")
}
\keyword{TreeTalkers}
