% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTT.R
\name{checkTT}
\alias{checkTT}
\title{Check TreeTalkers function}
\usage{
checkTT(dbTT, whichDay = Sys.Date() - 1)
}
\arguments{
\item{dbTT}{The ttDB returned by ReadServerData()}

\item{whichDay}{The day to filter data and produce reports? Defaults to Sys.Date()-1}
}
\value{
A list with the DB filtered on the selected wichDay and the alert DB with detected anomalies registered.
}
\description{
Finds anomalies and prepare data for reports.
}
\examples{
checkTT(TTdb, Sys.Date()-1)
}
\keyword{TreeTalkers}
