% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadNaturetalkers.R
\name{DownloadNaturetalkers}
\alias{DownloadNaturetalkers}
\title{Download function}
\usage{
DownloadNaturetalkers()
}
\value{

}
\description{
Appends all DBs at the specified server links in a single file ("TT_CloudDB.csv") which is updated with new records each time the TreeTalkersCheck package is executed.
}
\examples{
DownloadNaturetalkers()
}
\keyword{TreeTalkers}
