% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot}
\alias{Plot}
\title{Plot function}
\usage{
Plot(whichDay = Sys.Date() - 1)
}
\arguments{
\item{whichDay}{The day to filter data and produce reports? Defaults to Sys.Date()-1}
}
\value{
A list with the DB filtered on the selected wichDay and the alert DB with detected anomalies registered.
}
\description{
Calculates and saves several graphs into "C:/TreeTalkerDB".
}
\examples{
Plot("2020-09-16l")
}
\keyword{TreeTalkers}
