% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadServerData.R
\name{ReadServerData}
\alias{ReadServerData}
\title{Read server data function}
\usage{
ReadServerData()
}
\value{
A list with the cloudDB and the ttDB.
}
\description{
Adjusts and merges the 4B-4C and 4D-49 DBs and produces two files, one for TTs (ttDB.csv) and one for clouds (cloudDB.csv), both stand by for further analysis.
}
\examples{
ReadServerData()
}
\keyword{TreeTalkers}
