% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeTalkerCheck.R
\name{TreeTalkersCheck}
\alias{TreeTalkersCheck}
\title{The TreeTalkersCheck main function.}
\usage{
TreeTalkersCheck(whichDay = Sys.Date() - 1)
}
\arguments{
\item{whichDay}{The day to filter data and produce reports? Defaults to Sys.Date()-1}
}
\value{
This function saves graphs, DBs and reports into "C:/TreeTalkerDB".
}
\description{
This function allows you to execute the TreeTalkersCheck routine.
}
\examples{
TreeTalkersCheck()
TreeTalkersCheck("2020-09-16")
}
\keyword{TreeTalkers}
